requirements.txt
flask
python-dotenv
aiohttp
requests
openai
# langchain
# langchain_huggingface